More config tests

+++ util/file.go
@@ -96,11 +96,15 @@ func Close(f io.Closer) {
 }
 
 // RemoveFileAtPath removes a file at path (and any children) ignoring any error.
+// We do nothing if path == "".
 // This satisfies lint checks when using with defer and you don't care if there
 // is an error, so instead of:
 //   defer func() { _ = os.Remove(path) }()
 //   defer RemoveFileAtPath(path)
 func RemoveFileAtPath(path string) {
+	if path == "" {
+		return
+	}
 	_ = os.RemoveAll(path)
 }
